Refactored app to forward traffic to routers
